+++ sql/schema/stats/moveset_pokemon.sql
@@ -2,7 +2,7 @@ create table if not exists `moveset_pokemon`
 (
 `month` date not null,
 `format_id` tinyint unsigned not null,
-`pokemon_id` smallint unsigned not null,
+`pokemon_id` mediumint unsigned not null,
 
 `raw_count` mediumint unsigned not null,
 `viability_ceiling` tinyint unsigned null, # nullable
