+++ sql/schema/stats/usage_rated_pokemon.sql
@@ -5,7 +5,7 @@ create table if not exists `usage_rated_pokemon`
 `month` date not null,
 `format_id` tinyint unsigned not null,
 `rating` smallint unsigned not null,
-`pokemon_id` smallint unsigned not null,
+`pokemon_id` mediumint unsigned not null,
 
 `rank` smallint unsigned not null,
 `usage_percent` decimal(8, 5) not null,
